requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.4.0
Pillow>=10.0.0
numpy>=1.26.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
trimesh>=4.0.0
